Replace large script with new smaller scripts
